# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "BehaviorCount": {
        "description": "How many units of one behavior a book carries.",
        "properties": {
          "behavior": {
            "description": "Behavior label.",
            "title": "Behavior",
            "type": "string"
          },
          "units": {
            "description": "Unit count for this behavior.",
            "title": "Units",
            "type": "integer"
          }
        },
        "required": ["behavior", "units"],
        "title": "BehaviorCount",
        "type": "object"
      },
      "Book": {
        "description": "One book record (a single physical volume) exposed by the screens.",
        "properties": {
          "author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Author in romanized form.",
            "title": "Author"
          },
          "author_ar": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Author in Arabic, or None for anonymous and corporate texts.",
            "title": "Author Ar"
          },
          "blurb": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "One-sentence editorial summary, if any.",
            "title": "Blurb"
          },
          "canonical": {
            "anyOf": [
              {
                "enum": ["primary_reference", "primary", "secondary", "tertiary"],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Editorial rank, or None.",
            "title": "Canonical"
          },
          "category": {
            "description": "Category slug (matches Domain.categories[].slug).",
            "title": "Category",
            "type": "string"
          },
          "death_year_ah": {
            "anyOf": [
              {
                "maximum": 1500.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Author's death year, Hijri (unknown is None, never a sentinel).",
            "title": "Death Year Ah"
          },
          "death_year_ce": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Author's death year, CE.",
            "title": "Death Year Ce"
          },
          "foundational": {
            "default": false,
            "description": "True when the token \"foundational\" is among the source frontmatter tags.",
            "title": "Foundational",
            "type": "boolean"
          },
          "language": {
            "default": "Arabic",
            "description": "Primary language of the text.",
            "title": "Language",
            "type": "string"
          },
          "madhab": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Madhab when applicable.",
            "title": "Madhab"
          },
          "page_count": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Page count, if known.",
            "title": "Page Count"
          },
          "sect": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sect tagging (Imami, Sunni, etc.).",
            "title": "Sect"
          },
          "tags": {
            "description": "Tags parsed from the comma-separated frontmatter tag string.",
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "title_ar": {
            "description": "Arabic title.",
            "title": "Title Ar",
            "type": "string"
          },
          "title_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "English title, if available.",
            "title": "Title En"
          },
          "urn": {
            "description": "Opaque URN identifying the book across screens.",
            "title": "Urn",
            "type": "string"
          },
          "volume": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Volume number, if multi-volume.",
            "title": "Volume"
          }
        },
        "required": ["title_ar", "category", "urn"],
        "title": "Book",
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "description": "The body of every non-2xx response the API declares.",
        "properties": {
          "detail": {
            "description": "Human-readable failure description.",
            "title": "Detail",
            "type": "string"
          }
        },
        "required": ["detail"],
        "title": "ErrorEnvelope",
        "type": "object"
      },
      "ExtractionBookSummary": {
        "description": "Coverage summary for one book present in the manuscript artifact.",
        "properties": {
          "behaviors": {
            "description": "Unit counts per behavior, largest first.",
            "items": {
              "$ref": "#/components/schemas/BehaviorCount"
            },
            "title": "Behaviors",
            "type": "array"
          },
          "entities": {
            "description": "Total entities extracted for this book.",
            "title": "Entities",
            "type": "integer"
          },
          "first_page": {
            "description": "Lowest page on which a span starts.",
            "title": "First Page",
            "type": "integer"
          },
          "page_end": {
            "description": "Highest page any span reaches.",
            "title": "Page End",
            "type": "integer"
          },
          "pages_with_spans": {
            "description": "Distinct pages on which a span starts.",
            "title": "Pages With Spans",
            "type": "integer"
          },
          "spans": {
            "description": "Total spans extracted for this book.",
            "title": "Spans",
            "type": "integer"
          },
          "title_ar": {
            "description": "Arabic title from the catalog.",
            "title": "Title Ar",
            "type": "string"
          },
          "title_en": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "English title, if present.",
            "title": "Title En"
          },
          "units": {
            "description": "Total units extracted for this book.",
            "title": "Units",
            "type": "integer"
          },
          "urn": {
            "description": "Manifestation URN.",
            "title": "Urn",
            "type": "string"
          }
        },
        "required": [
          "urn",
          "title_ar",
          "first_page",
          "page_end",
          "pages_with_spans",
          "spans",
          "units",
          "entities",
          "behaviors"
        ],
        "title": "ExtractionBookSummary",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_Book_": {
        "properties": {
          "items": {
            "description": "Records in this slice.",
            "items": {
              "$ref": "#/components/schemas/Book"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "description": "Slice size requested.",
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Number of records skipped before this slice.",
            "minimum": 0.0,
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "description": "Total records matching the query, across all pages.",
            "minimum": 0.0,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": ["items", "total", "limit", "offset"],
        "title": "Page[Book]",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Knowledge-graph reader for classical Arabic manuscripts.",
    "title": "Shia Online Library",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/books": {
      "get": {
        "description": "Wrap the repo's (slice, total) into a Page[Book] envelope.",
        "operationId": "_list_books_api_books_get",
        "parameters": [
          {
            "description": "Filter by category slug.",
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by category slug.",
              "title": "Category"
            }
          },
          {
            "description": "Records per page.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 24,
              "description": "Records per page.",
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Records to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Records to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Book_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List books with pagination, optionally filtered by category slug.",
        "tags": ["books"]
      }
    },
...
    "/api/books/{urn}": {
      "get": {
        "description": "Return the book with the given URN, or raise ResourceNotFoundError.",
        "operationId": "get_book_api_books__urn__get",
        "parameters": [
          {
            "in": "path",
            "name": "urn",
            "required": true,
            "schema": {
              "title": "Urn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Book"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a single book by URN.",
        "tags": ["books"]
      }
    },
...
    "/api/books/{urn}/volumes": {
      "get": {
        "description": "Return every volume of the work containing ``urn``, ascending by volume\nnumber: the reader's volume switcher. A single-volume work returns just\nthat book. The membership comes from the same fold :func:`list_works`\nserves, so the two views can never disagree; a fold entry missing from the\nindex would be an invariant breach and raises KeyError loudly.",
        "operationId": "list_volumes_api_books__urn__volumes_get",
        "parameters": [
          {
            "in": "path",
            "name": "urn",
            "required": true,
            "schema": {
              "title": "Urn",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Book"
                  },
                  "title": "Response List Volumes Api Books  Urn  Volumes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List every volume of the work containing this book, ascending.",
        "tags": ["books"]
      }
    },
    "/api/citations/{surah}/{aya}/books": {
      "get": {
        "description": "Reverse concordance: how many distinct books cite ``surah:aya``.",
        "operationId": "_books_citing_api_citations__surah___aya__books_get",
        "parameters": [
          {
            "in": "path",
            "name": "surah",
            "required": true,
            "schema": {
              "title": "Surah",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "aya",
            "required": true,
            "schema": {
              "title": "Aya",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response  Books Citing Api Citations  Surah   Aya  Books Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "The requested resource does not exist (unknown urn, id, surah, or ayah)."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Count of distinct books citing a given surah:aya (reverse concordance).",
        "tags": ["citations"]
      }
    },
...
    "/api/dev/extraction/books": {
      "get": {
        "description": "Return one coverage summary per book in the artifact, URN-ordered.\n\nTitles come from the catalog: a URN present in the artifact but absent\nfrom the catalog raises ResourceNotFoundError \u2014 a stale artifact must be\nrebuilt, not partially listed.",
        "operationId": "extraction_summaries_api_dev_extraction_books_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExtractionBookSummary"
                  },
                  "title": "Response Extraction Summaries Api Dev Extraction Books Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Dev tools are disabled on this deployment (SOL_DEV_TOOLS unset); declared so generated client types represent what callers actually see."
          }
        },
        "summary": "List the books in the manuscript artifact with extraction coverage.",
        "tags": ["extraction"]
      }
    },
...
  }
}